Extract indent_c.c from misc1.c
